requirements.txt
asgiref==3.3.4
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2.2
django-memcache-status==2.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
entrypoints==0.3
flake8==3.7.9
gunicorn==20.1.0
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.17.1
python-memcached==1.59
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.2.0
